+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 ## [Unreleased][]
 
 [Unreleased]: https://github.com/chaostoolkit-incubator/chaostoolkit-aws/compare/0.17.0...HEA
-### Added
 
+### Added
+- added new variable for cloudwatch probes `get_metric_statistics` and `get_metric_data` called `dimensions` to allow for multiple dimension queries
 * Added `.github/workflows/check_pr.yaml` which checks if a PR has modified the
 CHANGELOG.md and if it changed/added tests